+++ src/OutreachFlow.Application/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using OutreachFlow.Application.Contacts;
+using OutreachFlow.Application.ContactActivities;
 using OutreachFlow.Application.EmailDrafts;
 using OutreachFlow.Application.EmailTemplates;
 using OutreachFlow.Application.Organizations;
@@ -17,6 +18,7 @@ public static class ServiceCollectionExtensions
         services.AddScoped<IContactService, ContactService>();
         services.AddScoped<IOrganizationService, OrganizationService>();
         services.AddScoped<ITagService, TagService>();
+        services.AddScoped<IContactActivityService, ContactActivityService>();
         services.AddScoped<ISenderProfileService, SenderProfileService>();
         services.AddScoped<IEmailTemplateService, EmailTemplateService>();
         services.AddScoped<IAttachmentAssetService, AttachmentAssetService>();